+++ CHANGELOG.md
@@ -5,7 +5,46 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 0.0.1 - 2025-10-26
+<!-- ## [Unreleased] - 2025-10-26
+
+[Unreleased]: https://github.com/nirvikpurkait/tsp-temp/compare/v0.0.2...Unreleased -->
+
+## [0.0.3] - 2025-10-26
+
+[0.0.3]: https://github.com/nirvikpurkait/tsp-temp/compare/v0.0.2...0.0.3
+
+### Added
+
+- `--template-branch` option into cli to provide a remote repositiry branch
+- `--expensive-way` option into cli to provide support for other remote perository provider other than **GitHub** and **GitLab**. (The options describes itself it is an expensive way to generate a template, such as consumes more data)
+
+### Removed
+
+- Un-necessary dependencies
+
+### Optimized
+
+- Template creation process from remote repository with **GitHub** and **GitLab** (using less data)
+
+<!-- ________________________________________________ -->
+
+---
+
+## [0.0.2] - 2025-10-26
+
+[0.0.2]: https://github.com/nirvikpurkait/tsp-temp/compare/v0.0.1...v0.0.2
+
+### Fixed
+
+- Binary files of `tsx-temp` was not including into npm.
+
+## <!-- ________________________________________________ -->
+
+---
+
+## [0.0.1] - 2025-10-26
+
+[0.0.1]: https://github.com/nirvikpurkait/tsp-temp/compare/v0.0.1
 
 ### Added